Repair NQueens.
At Berkeley, CS61C students are taught not to declare counters in "for" loops because it used to be UB/forbidden in older versions of C. The best solution to this is probably not to use a global variable (!!!!!!!) as a counter (!!!!!), especially in a program with parallelism.

This change also gives the program a more explicit error message if the output is incorrect.